+++ backend/models/visit.go
@@ -9,5 +9,5 @@ type Visit struct {
 	CheckpointID int        `json:"checkpoint_id"`    // Идентификатор чекпоинта (числовой, автоинкремент).
 	StartAt      time.Time  `json:"start_at"`         // Время начала визита (входа).
 	EndAt        *time.Time `json:"end_at,omitempty"` // Время окончания визита; если визит активен, это поле будет nil.
-	Duration     float64    `json:"duration"`         // Длительность визита в секундах (вычисляется при завершении визита).
+	Duration     int        `json:"duration"`         // Длительность визита в секундах (вычисляется при завершении визита).
 }
